Further symmetry changes (by ljkiernan1)
